Refine app architecture
Framework for displaying information should be settled, and can be
extended to any other features we might require.

Model classes are solely for storing data and functionality, and any rendering
must be handled by a corresponding `View` class.
* All `View` children classes inherit `subviews` and `addSubview(View)`, which
  act as components that can be rendered within the superview.
* `PageView` renders the default template for any given page of the
  application.
* `ListView` and `ItemView` work in tandem to generate units of rows.
  * If needed, we could increase the complexity of `ItemView` to create
    more complex lists.

On launch, application data is expected to be loaded into the `AppData`
singleton, which all other classes can then read from.