Implement a catch-all ListView
Use delegation and generics to adapt `NumberedActionListView` into
`ListView` and make it more adaptable.

Instead of creating a new subclass of `ListView` whenever we need to
pass in custom rendering of its items, we can instead have our item
classes implement the `ItemViewDataSource` interface, and implement the
`render` function.

Only issue is that due to a lack of class extension features, we will
have to implement the `render` function within the model class itself.